Update Commands.h
I was trying to add an extended option to auto host by waiting for the whole timer to run out, 3mins but my coding skills are not up to it.
So I guess this is a feature request unless I can figure it out myself then I will but a proper pull request in.

If others join the raid then the game auto readies the player, no extra button pushing. If the lobby is empty then it doesn't start and it would not break the bot when it tries the friend request method of disconnecting.